+++ packages/integrations/src/tracing.ts
@@ -1,11 +1,11 @@
 import { EventProcessor, Hub, Integration } from '@sentry/types';
-import { consoleSandbox, fill, getGlobalObject, isMatchingPattern, supportsNativeFetch } from '@sentry/utils';
+import { fill, getGlobalObject, isMatchingPattern, logger, supportsNativeFetch } from '@sentry/utils';
 
 /** JSDoc */
 interface TracingOptions {
   tracingOrigins?: Array<string | RegExp>;
-  traceXHR?: boolean;
   traceFetch?: boolean;
+  traceXHR?: boolean;
   autoStartOnDomReady?: boolean;
 }
 
@@ -37,16 +37,12 @@ export class Tracing implements Integration {
    */
   public constructor(private readonly _options: TracingOptions = {}) {
     if (!Array.isArray(_options.tracingOrigins) || _options.tracingOrigins.length === 0) {
-      consoleSandbox(() => {
-        const defaultTracingOrigins = ['localhost', /^\//];
-        // @ts-ignore
-        console.warn(
-          'Sentry: You need to define `tracingOrigins` in the options. Set an array of urls or patterns to trace.',
-        );
-        // @ts-ignore
-        console.warn(`Sentry: We added a reasonable default for you: ${defaultTracingOrigins}`);
-        _options.tracingOrigins = defaultTracingOrigins;
-      });
+      const defaultTracingOrigins = ['localhost', /^\//];
+      logger.warn(
+        'Sentry: You need to define `tracingOrigins` in the options. Set an array of urls or patterns to trace.',
+      );
+      logger.warn(`Sentry: We added a reasonable default for you: ${defaultTracingOrigins}`);
+      _options.tracingOrigins = defaultTracingOrigins;
     }
   }
 
@@ -151,7 +147,7 @@ export class Tracing implements Integration {
         const self = getCurrentHub().getIntegration(Tracing);
         if (self && self._options.tracingOrigins) {
           const url = args[0] as string;
-          const options = args[1] as { [key: string]: any };
+          const options = (args[1] = (args[1] as { [key: string]: any }) || {});
 
           let whiteListed = false;
           self._options.tracingOrigins.forEach((whiteListUrl: string | RegExp) => {
@@ -160,7 +156,7 @@ export class Tracing implements Integration {
             }
           });
 
-          if (options && whiteListed) {
+          if (whiteListed) {
             if (options.headers) {
               if (Array.isArray(options.headers)) {
                 options.headers = [...options.headers, ...Object.entries(getCurrentHub().traceHeaders())];