+++ packages/types/src/options.ts
@@ -66,7 +66,10 @@ export interface Options {
   /** Sets the distribution for all events */
   dist?: string;
 
-  /** The maximum number of breadcrumbs sent with events. Defaults to 100. */
+  /**
+   * The maximum number of breadcrumbs sent with events. Defaults to 30.
+   * Values over 100 will be ignored and 100 used instead.
+   */
   maxBreadcrumbs?: number;
 
   /** Console logging verbosity for the SDK Client. */