+++ packages/utils/src/error.ts
@@ -1,3 +1,5 @@
+import { setPrototypeOf } from './polyfill';
+
 /** An error emitted by Sentry SDKs and related utilities. */
 export class SentryError extends Error {
   /** Display name of this error instance. */
@@ -8,6 +10,6 @@ export class SentryError extends Error {
 
     // tslint:disable:no-unsafe-any
     this.name = new.target.prototype.constructor.name;
-    Object.setPrototypeOf(this, new.target.prototype);
+    setPrototypeOf(this, new.target.prototype);
   }
 }